+++ lib/Cookie.php
@@ -2,7 +2,7 @@
 
 namespace Facebook\WebDriver;
 
-use InvalidArgumentException;
+use Facebook\WebDriver\Exception\Internal\LogicException;
 
 /**
  * Set values of an cookie.
@@ -36,10 +36,10 @@ class Cookie implements \ArrayAccess
     public static function createFromArray(array $cookieArray)
     {
         if (!isset($cookieArray['name'])) {
-            throw new InvalidArgumentException('Cookie name should be set');
+            throw LogicException::forError('Cookie name should be set');
         }
         if (!isset($cookieArray['value'])) {
-            throw new InvalidArgumentException('Cookie value should be set');
+            throw LogicException::forError('Cookie value should be set');
         }
         $cookie = new self($cookieArray['name'], $cookieArray['value']);
 
@@ -107,7 +107,7 @@ class Cookie implements \ArrayAccess
     public function setDomain($domain)
     {
         if (mb_strpos($domain, ':') !== false) {
-            throw new InvalidArgumentException(sprintf('Cookie domain "%s" should not contain a port', $domain));
+            throw LogicException::forError(sprintf('Cookie domain "%s" should not contain a port', $domain));
         }
 
         $this->offsetSet('domain', $domain);
@@ -258,11 +258,11 @@ class Cookie implements \ArrayAccess
     protected function validateCookieName($name)
     {
         if ($name === null || $name === '') {
-            throw new InvalidArgumentException('Cookie name should be non-empty');
+            throw LogicException::forError('Cookie name should be non-empty');
         }
 
         if (mb_strpos($name, ';') !== false) {
-            throw new InvalidArgumentException('Cookie name should not contain a ";"');
+            throw LogicException::forError('Cookie name should not contain a ";"');
         }
     }
 
@@ -272,7 +272,7 @@ class Cookie implements \ArrayAccess
     protected function validateCookieValue($value)
     {
         if ($value === null) {
-            throw new InvalidArgumentException('Cookie value is required when setting a cookie');
+            throw LogicException::forError('Cookie value is required when setting a cookie');
         }
     }
 }